Update from warnings to errors
tweak files to force errors rather than simple warnings.

+++ WinUserProfile.ps1
@@ -5,22 +5,30 @@ $collection = Get-CimInstance -ClassName Win32_UserProfile -ComputerName $machin
 $active = Get-WmiObject -Class Win32_ComputerSystem -computername $machine | Select-Object UserName
 #clean up the active user information
 $name = $($active.UserName -split '\\')[1]
-Write-Host "$name is currently logged in"
+Write-Warning -Message "$name is currently logged in"
 #Start the selection menu
 [INT]$profile = '0'
 foreach ($item in $collection) {
     $opt = ($item.localpath -split '\\')[2]
     If ($name -ne $opt) {
-        Write-Host $profile $opt
+        Write-Output "$profile $opt"
     }else{
-        Write-Host "Cannot select $name - Active User"
+        Write-Warning -Message "Cannot select $name - Active User"
         $disallowed = $profile
     }
     $profile ++
 }
+$profile --
 [int]$selection = Read-Host "Make a selection"
 #Prevent active user from selection, then remove identified user
-If ($selection -ne $disallowed){
-    Write-Host "Selecting user $($collection[$selection].LocalPath) in $machine for removal"
-    Remove-CimInstance -computername $machine $collection[$selection] -Confirm
+if ($selection -le $profile) {
+    
+    If ($selection -ne $disallowed){
+        Write-Output -Message "Selecting user $($collection[$selection].LocalPath) in $machine for removal"
+        Remove-CimInstance -computername $machine $collection[$selection] -Confirm
+    } Else {
+        Write-Error -Message "Cannot delete Active user"
+    }
+} Else {
+    Write-Error -Message "$selection was not a valid input. Options go to $profile."
 }
